Design fix
Because of the way faults work, if you want to, say, extended the indexed
address modes, we need some way to pass back the new effective address.  So
I'm moving one of the internal CPU fields to a documented field, and changing
the name to futher reflect what it is.  So the 'addr' field is changed to the
'ea' field.  Extentions and new addressing modes need to set this field.